tests are now running fine
