+++ package.json
@@ -17,6 +17,7 @@
         "build:navigation": "turbo run build --filter @equinor/mad-navigation",
         "dev:insights": "turbo run dev --filter @equinor/mad-insights",
         "build:insights": "turbo run build --filter @equinor/mad-insights",
+        "test:insights": "turbo run test --filter @equinor/mad-insights",
         "dev:chronicles": "turbo run dev --filter @equinor/mad-chronicles",
         "build:chronicles": "turbo run build --filter @equinor/mad-chronicles",
         "build:maintenance-api": "turbo run build --filter @equinor/mad-maintenance-api-ts-wrapper",
